requirements.txt
Flask==3.0.3
Werkzeug==3.0.2
requests==2.31.0
python-dotenv==1.0.1
elasticsearch==8.13.0
PyPDF2==3.0.1
python-slugify==8.0.4
gunicorn==21.2.0
